+++ EmailWebService/Interfaces/IEmailDbROController.cs
@@ -2,9 +2,9 @@
 {
     public interface IEmailDbROController
     {
-        long GetIdentityCodeId(string IdentityCode);
-        IAppPermisionModel GetAppPermision(long IdentityCodeId, string ServiceName);
-        IEmailConfigurationModel GetEmailConfiguration(long IdentityCodeId);
-        string GetEmailBody(long IdentityCodesId, string SchemaName, List<(string Name, string Value)> VariablesList);
+        int GetIdentityCodeId(string IdentityCode);
+        IAppPermisionModel GetAppPermision(int IdentityCodeId, string ServiceName);
+        IEmailConfigurationModel GetEmailConfiguration(int IdentityCodeId);
+        string GetEmailBody(int IdentityCodesId, string SchemaName, List<(string Name, string Value)> VariablesList);
     }
 }
