+++ EmailWebService/Models/AppEmailServiceSettingsDbModel.cs
@@ -6,11 +6,11 @@ namespace EmailWebService.Models
     public class AppEmailServiceSettingsDbModel : IAppEmailServiceSettingsDbModel
     {
         [Key]
-        public long Id { get; set; }
+        public int Id { get; set; }
         [Required]
-        public long IdentityCodeId { get; set; }
+        public int IdentityCodeId { get; set; }
         [Required]
-        public long EmailConfigurationId { get; set; }
+        public int EmailConfigurationId { get; set; }
         public IdentityCodeDbModel IdentityCode { get; set; }
         public EmailConfigurationDbModel EmailConfiguration { get; set; }
     }
