+++ EmailWebService/Models/AppPermisionDbModel.cs
@@ -6,9 +6,9 @@ namespace EmailWebService.Models
     public class AppPermisionDbModel : IAppPermisionDbModel
     {
         [Key]
-        public long Id { get; set; }
+        public int Id { get; set; }
         [Required]
-        public long  IdentityCodeId { get; set; }
+        public int IdentityCodeId { get; set; }
         [Required]
         public string ServiceName { get; set; }
         public IdentityCodeDbModel IdentityCode { get; set; }
