+++ EmailWebService/Models/AppPermisionModel.cs
@@ -4,8 +4,8 @@ namespace EmailWebService.Models
 {
     public class AppPermisionModel : IAppPermisionModel
     {
-        public long Id { get; set; }
-        public long IdentityCodeId { get; set; }
+        public int Id { get; set; }
+        public int IdentityCodeId { get; set; }
         public string ServiceName { get; set; }
     }
 }
